Get Wan 2.1 working on the Nvidia DGX Spark
PyTorch upstream doesn't (as of December 2025) provide an official
container that supports aarch64, and manually building one with CUDA
support included proved to be difficult.

Luckily Nvidia provides a base container with both support for aarch64
and the (very new) Nvidia GB20B GPU found in their Nvidia GB10 chip, so
I've switched to that base container.

I ran two full runs of the wan21 container; one with the recommended
settings and one with the offload tuneable set to False. The runs both
ran at around 13.7 s/it, took 21 and 16 mins respectively, and used
~14 GB of CPU RAM and ~9 GB and ~19 GB of GPU RAM respectively.